requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
pyarrow>=14.0.0
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.10.0
google-cloud-run>=0.10.0
google-auth>=2.23.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
requests>=2.31.0
matplotlib>=3.7.0

